+++ test/WebApi20.FunctionalTests/Controllers/ValuesController.cs
@@ -3,12 +3,20 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.Extensions.Logging;
 
 namespace WebApi20.FunctionalTests.Controllers
 {
     [Route("api/[controller]")]
     public class ValuesController : Controller
     {
+        ILogger<ValuesController> logger;
+
+        public ValuesController(ILogger<ValuesController> logger)
+        {
+            this.logger = logger;
+        }
+
         // GET api/values
         [HttpGet]
         public IEnumerable<string> Get()
@@ -20,6 +28,10 @@ namespace WebApi20.FunctionalTests.Controllers
         [HttpGet("{id}")]
         public string Get(int id)
         {
+            logger.LogError("error logged");
+            logger.LogWarning("warning logged");
+            logger.LogInformation("information logged");
+            logger.LogTrace("trace logged");
             return "value";
         }